requirements.txt
face_detection==0.2.1
opencv_python==4.5.5.64
Pillow==9.0.1
streamlit==1.7.0
